Move IsValidEntryPointPath
