+++ src/Cli/dotnet/Commands/Run/VirtualProjectBuildingCommand.cs
@@ -1091,32 +1091,6 @@ internal sealed class VirtualProjectBuildingCommand : CommandBase
         }
     }
 
-    public static bool IsValidEntryPointPath(string entryPointFilePath)
-    {
-        if (!File.Exists(entryPointFilePath))
-        {
-            return false;
-        }
-
-        if (entryPointFilePath.EndsWith(".cs", StringComparison.OrdinalIgnoreCase))
-        {
-            return true;
-        }
-
-        // Check if the first two characters are #!
-        try
-        {
-            using var stream = File.OpenRead(entryPointFilePath);
-            int first = stream.ReadByte();
-            int second = stream.ReadByte();
-            return first == '#' && second == '!';
-        }
-        catch
-        {
-            return false;
-        }
-    }
-
     public static readonly ErrorReporter ThrowingReporter =
         static (sourceFile, textSpan, message) => throw new GracefulException($"{sourceFile.GetLocationString(textSpan)}: {FileBasedProgramsResources.DirectiveError}: {message}");
 }
